Refactored OpenAPI and Structured Function tools, Strands Observability
