+++ src/bedrock_agentcore_starter_toolkit/services/import_agent/scripts/__init__.py
@@ -1,1 +1,1 @@
-"""Translation from Bedrock Agents to Langchain/Strands + Genesis Agents."""
+"""Translation from Bedrock Agents to Langchain/Strands + AgentCore Agents."""
